Use signed integer type instead of size_t

+++ src/Sound/AudioEmitter.h
@@ -16,7 +16,7 @@
 namespace openblack::audio
 {
 
-using AudioEmitterId = size_t;
+using AudioEmitterId = int64_t;
 using AudioSourceId = unsigned int;
 using AudioBufferId = unsigned int;
